requirements.txt
streamlit==1.46.0
spacy==3.8.7
transformers==4.52.4
torch
pdfminer.six==20231228  # Downgraded to match pdfplumber
python-docx==1.2.0
pytesseract==0.3.13
pdf2image==1.17.0
pillow==11.2.1
keybert==0.7.0
pymupdf==1.24.4
pdfplumber==0.11.0
nltk==3.8.1
